split admin/non-admin code
